% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_ana_logrank.R
\name{logrank_pv}
\alias{logrank_pv}
\title{Geting P value using logrank from a dataframe with multiple varnames}
\usage{
logrank_pv(surv.data, time, vital, group, positive.lab, factor.Levels)
}
\arguments{
\item{surv.data}{a dataframe with multiple varnames: column: time, vital status, var names; row: samples}

\item{time}{colnames of survial time}

\item{vital}{colnames of vital status}

\item{group}{colnames  var names as colnames}

\item{positive.lab}{positive label defines event occurs; such as 'Dead'}

\item{factor.Levels}{define factor levels of the var name (group)}
}
\value{
a vector of var name and pvalue
}
\description{
Geting P value using logrank from a dataframe with multiple varnames
}
\examples{
logrank_pv(surv.data = sur_data, time = 'days_to_last_follow_up', vital = 'vital_status', group = snp,
                  positive.lab = 'Dead', factor.Levels = c('wild', 'mutation'))
}
