% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_cutoff.R
\name{surv_bestcutoff}
\alias{surv_bestcutoff}
\title{Find the best cutoff threhold for a numeric var; if median grouping could not give a significant P value}
\usage{
surv_bestcutoff(
  surv.data,
  time,
  vital,
  numeric.varName,
  positive.lab,
  minprop = 0.3
)
}
\arguments{
\item{surv.data}{a dataframe with multiple varnames: column: time, vital status, var names; row: samples}

\item{time}{colnames of survial time}

\item{vital}{colnames of vital status}

\item{numeric.varName}{colnames  var names as colnames (must be numeric vectors)}

\item{positive.lab}{positive label defines event occurs; such as 'Dead'}

\item{minprop}{minimal proption to prevent unbanlanced grouping}
}
\value{
a best cutoff surv fit cutoff value
}
\description{
Find the best cutoff threhold for a numeric var; if median grouping could not give a significant P value
}
\examples{
surv_bestcutoff(surv.data = sur_data, time = "days_to_last_follow_up", vital = "vital_status", numeric.varName = "total_perMB", positive.lab = 'Dead')
}
