% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_ploting.R
\name{surv_plot1}
\alias{surv_plot1}
\title{Kaplan Meier plot using survminer package}
\usage{
surv_plot1(
  surv.data,
  time,
  vital,
  group,
  positive.lab,
  factor.Levels,
  title = "",
  x.lab = "Time",
  y.lab = "Survival probability",
  font.x = 15,
  font.y = 15,
  font.tickslab = 15,
  font.legend = 12,
  risk.table = F,
  conf.int = F,
  col = (ggsci::pal_lancet(alpha = 0.8))(9)
)
}
\arguments{
\item{surv.data}{a dataframe with multiple varnames: column: time, vital status, var names; row: samples}

\item{time}{colnames of survial time}

\item{vital}{colnames of vital status}

\item{group}{colnames  var names as colnames}

\item{positive.lab}{positive label defines event occurs; such as 'Dead'}

\item{factor.Levels}{define factor levels of the var name (group)}

\item{title}{title of the plot}

\item{x.lab}{x lab}

\item{y.lab}{y lab}

\item{font.x}{font size of x axis title, default 15}

\item{font.y}{font size of y axis title,default 15}

\item{font.tickslab}{font size of labs in x and y,default 15}

\item{font.legend}{font size of legend title and labs, default 12}

\item{risk.table}{if show risk table, default False}

\item{conf.int}{if add confidence ,default False}

\item{col}{palette ,defalut lancet}
}
\value{
a KM plot
}
\description{
Kaplan Meier plot using survminer package
}
\examples{
surv_plot1(surv.data = sur_data, time = 'days_to_last_follow_up', vital = 'vital_status', group = snp,
                     positive.lab = 'Dead', factor.Levels = c('wild', 'mutation'))

}
