% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_ploting.R
\name{surv_plot2}
\alias{surv_plot2}
\title{Custom version of Kaplan Meier plot from IM package}
\usage{
surv_plot2(
  surv.data,
  time,
  vital,
  group,
  positive.lab,
  factor.Levels,
  main = "",
  xLab = "OS (months)",
  yLab = "Probability of survival",
  cols,
  ltypes = 1,
  plotMedian = FALSE,
  pval = c("logrank", "coxph", "none"),
  legPos = "topright",
  mar = c(12, 9, 3, 2),
  plot.nrisk = FALSE,
  nrisk.interval = 2,
  timemark = TRUE,
  lwd = 2,
  cexMedian = 0.8,
  cexLegend = 0.9,
  ...
)
}
\arguments{
\item{surv.data}{a dataframe with multiple varnames: column: time, vital status, var names; row: samples}

\item{time}{colnames of survial time}

\item{vital}{colnames of vital status}

\item{group}{colnames  var names as colnames}

\item{positive.lab}{positive label defines event occurs; such as 'Dead'}

\item{factor.Levels}{define factor levels of the var name (group)}

\item{main}{main character; main title for plot}

\item{xLab}{character; x-axis label for plot}

\item{yLab}{character; y-axis label for plot}

\item{cols}{character vector; colors, default to darkgreen, darkmagenta, cyan4, darkorange, darkred}

\item{ltypes}{numeric vector; line types (lty) for plot}

\item{plotMedian}{logical; show the median survival for each group be plotted, default False}

\item{pval}{type of p-value computed, either "logrank" or "coxph" (the latter can only be computed on two groups)
or "none" don't show}

\item{legPos}{legend position for legend , default "topright"}

\item{mar}{plotting margins, default c(12,9,3,2)}

\item{plot.nrisk}{logical; show number of samples for each group and time point,
be plotted underneath graph .default False}

\item{nrisk.interval}{numeric; spacing of time intervals for which number of samples
is given; defaults to 2}

\item{timemark}{logical; if set to TRUE (default), censoring marks are added to survival curves}

\item{lwd}{line width for survival curves, default 2}

\item{cexMedian}{numeric; relative size of font for median survival times , default 0.8}

\item{cexLegend}{numeric; relative size of legend font, default 0.9}

\item{...}{other arguments passed to plot.survfit}
}
\value{
a KM plot
}
\description{
Custom display of survival data, including plotting lines for median survival and adding number at risk
}
